Stop using a queue for logging stuff to a terminal.
Use `slog_async` will drop messages on the floor, if too many messages are
sent at once. This can happen when processing the output of `git checkout`
while updating lists for example. So use a `Mutex` instead of
`slog-async` to not drop any messages. Since the command-processing is
single threaded anyway, there shouldn't be any contention.