remove automatic blacklist
The automatic blacklist was added a while back and it included all
crates that consistently failed for multiple past runs. It hasn't been
updated since then and maintaining it is a pain. Also we're losing
coverage since the crates in the blacklist might be fixed in the future.

This removes all the automatic entries from the blacklist.