Log backtraces captured by `failure::Error`
Previously, `AsFail::as_fail` was being called on `failure::Error`
instances passed to `report_failure`. This extracted the underlying
error object, but discarded the backtrace stored when the
`failure::Error` was constructed.

This PR adds a `HasBacktrace` trait to extract the backtrace from from
the various `failure::Error` and `failure::Context` instances that are
passed to `report_failure`. It also prints a message reminding the user
to enable backtraces when an error occurs.