agent: purge build directories only on new experiments
Before this commit, the agent purged the build directories every time an
experiment was started. This was fine before distributed experiments
were introduced, but it clashed with them as distributed experiments are
implemented as tons of little sub-experiments.

Because of that the build directories were purged every 32 crates,
hurting build caches a lot. This commit changes the agent to purge build
directories both when it's started and when starting to test an
experiment with a name different than the previous one.