server: allow retrying reports when they're stuck in progress
We had a Crater issue in the weekend where report generation was stuck.
To fix it I had to manually log into the server and change the status of
the experiment to report-failed, allowing retry-report to work. This
commit changes the retry-report command to allow retrying stuck reports
as well, removing the need to tweak the database.