Remove stale recommendation about Defender
We fixed the performance problem installing docs last year when we overhauled rustup's performance and got Microsoft's Defender team to look into the Defender behaviour vis-a-vis Rustup. It shouldn't be a problem now, and the bug is closed.

Also we now support a minimal profile for automated intallations where no human is present, so adding that seems useful to further speed things up.