Use second parent of merge commit when detecting outdated 'try' build
In https://github.com/rust-lang/crater/pull/540, I used the `build_sha`
field to detect when the current PR head does not match the PR head used
when creating the 'try' build. However, 'build_sha' corresponds to the
*first* parent of the merge commit, which corresponds to the commit on
the rust-lang/rust 'master' branch.

We now retrieve the parents of the merge commit from GitHub, and check
the first parent against the current head commit of the PR.

During local testing, I had created a merge commit with the order of the
parents reversed, so the check appeared to work fine for me.