Retain 5 GB of slack space, rather than 15% of the disk
On a 2 TB disk, 15% is 300 GB: way more than necessary for some margin of
safety. In general, on large disks, this is always going to be the case, whereas
on smaller disks 15% slack space may be pretty small. Specify free space as a
constant number of gigabytes instead.

In theory it might be better to have it be worker-dependent (e.g., 1 GB per
worker) but that seems more complicated and likely not really necessary.