Add a CPU limit, raise example memory limit
The combination of no job limit and a low memory limit has been causing
a lot of OOMs in crater runs. This adds a job limit and enforces it
including for doctests, which are built by the test runner. If nothing
else, setting this CPU limit provides reproducible OOMs because many
crates have peak memory usage proportional to build parallelism.