Recover gracefully from agent failure
Previously, any agent reporting an error would cause an experiment to come to a
hard stop. This is noisy (leads to a user-visible report) and, at least lately,
not actually representative of a terminal failure. Instead, we take the approach
of re-queueing any crates assigned to that agent and logging an error on the
server host specifying which agent failed.

As the error message is now not typically posted publicly (and no longer needs
to be short for being in a GitHub comment), we will also have more opportunity
to add extra metadata over time to that message (in future PRs) enabling
diagnosing and fixing these errors.

In the future, we will likely also want some tracking/reporting of these
failures such that hard errors across all agents lead to alerting the infra
team, but this may be best done through Prometheus metrics rather than GitHub
"end-user" visible comments (which are likely not the best way to reach infra
team members regardless).