Remove mutex lock around record_progress endpoint
See comment added in source for why this is OK. We are trying to reduce the
runtime of this endpoint, as the longer it takes the fewer results we can
receive. Obviously, the agents submitting results are the true bound on
performance, but if the server storing results is near capacity (as we currently
are, on a single thread), then we may be artificially reducing our throughput.
Certainly, the endpoint performance is a bound on the *possible* highest level
of throughput with an infinitely fast agent pool.