Drop stub needs_exec
Runners did not keep track of any state (or support querying upstream) for
completion; we expect individual tasks to only be sent to runners if they should
attempt to run them, and if duplicate work is done that's OK (each individual
task is cheap).

This function already always returned 'true' so just delete it and propagate the
effects of that.