Drop result kinds from experiment page
This is a pretty expensive query to compute, and while that could be optimized
with some amount of indexing and creative SQL, it's not clear that this table
delivers much value to the user. If we wanted to present partial reports, that
could be done on a periodic basis (e.g., creating the index.html page every
hour) and would seem largely more useful.

In the meantime, it's better that the UI loads quickly than that we deliver this
data.