Switch crate queue to retrieve crates based on timeout
Previously, a given crate was assigned to a particular agent, and then that
agent had to either complete or fail (e.g., panic) before it was 'released' for
another agent to attempt. This is error prone, and makes it harder for agents to
be ephemeral (including their name), as it means that it requires manual
intervention if an agent just vanishes to put the assigned crates back into the
general queue.

We switch to a different assignment scheme. Now, when an agent requests crates
from a given experiment, we will dequeue crates that are queued (i.e.,
incomplete) and that have not started yet **or** have started >20 minutes ago.
The current agent timeout is blanket set at 15 minutes, and even if it was
longer, we expect that most crates build quite quickly. Duplicate builds are
also fine for our case (we'll just get a couple extra results but that'll be
deduplicated already when inserting into that table).