Rework getting next crate to be per-worker, rather than a global queue
The previous commit added the notion of timeouts on retrieved crates from the
server, which meant that when an agent (which has multiple workers) requested
the 1024 crates from the server, the tail of that request would likely have
timed out by the time the agent started/completed working on those crates. This
isn't technically a problem, but it is likely to lead to duplicated work for
that tail, so we avoid it by retrieving crates one at a time with the new
commit.