Increase retry intervals
With retries in a 5 second interval and our current peak of ~424 worker threads,
we end up with 85 requests/second even fully uniformly distributed, which is
unsustainable. (Note that new work is being added at *roughly* ~8 requests per
second from rustc jobs completing, so we may not be able to drain the queue at
all since our processing takes ~150ms at p99).

Spreading across eight minutes brings us to ~1 rps of retry load.  In practice
it'll be higher since any workers *not* seeing errors will return to the
previous compute, but that brings us to roughly ~10 rps, which we should
hopefully largely be able to handle, given that our p50 is only 60ms).