Distinct build directories per toolchain
We see a number of failures from build scripts or tests failing when run
in the same directory. These are 99% likely to be bugs in that crate,
but Crater's goal isn't to find bugs in crates but bugs/regressions in
rustc, so we should try to avoid sharing a build directory as we did
before.

In practice this additional level of nesting should be minimally
impactful in terms of cache hits, as caching across toolchains wasn't
possible anyway. It may even improve things.

Our purging algorithm is the same as before.