Refactor to preserve logs throughout all retries
This moves all pushes of logs to the central server out of the execution
of tasks and into the worker function, which ensures that we don't
overwrite previous results for the same crate multiple times (when
retrying its run). That should hopefully preserve the logs as well
throughout the retries.