Reduce scans of experiment_crates to find next crate
These scans are relatively expensive, so instead we internally cache a
small amount of crates. The expectation is that this should allow us to
scan ~once rather than repeatedly re-scanning as time goes by. I
experimented with a partial index but didn't get any meaningful results
(possibly made things worse), it's not super clear *why*, but the hope
is that this is more or less just an improvement.