seperate prepare failure from broken/errored crate
Prepare failure are likely spurious and as such are included in the retry-regressed-list.txt to give the crates another change in a retry.
This is different to a crate being Broken which is probably non-spurious.

Iif either base or test toolchain is a prepare failure the runs can't give any insight as we are missing either a base or a test result.